requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
streamlit==1.31.1
langchain-google-genai==1.0.1
google-generativeai==0.4.1
langchain-core>=0.1.42,<0.2.0
langchain-text-splitters==0.0.1
faiss-cpu==1.7.4
langchain-community==0.0.29
python-dotenv==1.0.1
numpy>=1.20.0
